requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-dotenv
openai
python-multipart
